refactor pipelines
