+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentGraphPipeline.kt
@@ -1,23 +1,10 @@
 @file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-@file:OptIn(InternalAgentsApi::class)
 
 package ai.koog.agents.core.feature.pipeline
 
 import ai.koog.agents.core.agent.config.AIAgentConfig
-import ai.koog.agents.core.agent.context.AIAgentGraphContextBase
-import ai.koog.agents.core.agent.entity.AIAgentNodeBase
-import ai.koog.agents.core.agent.entity.AIAgentSubgraphBase
-import ai.koog.agents.core.agent.execution.AgentExecutionInfo
-import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.feature.AIAgentGraphFeature
 import ai.koog.agents.core.feature.config.FeatureConfig
-import ai.koog.agents.core.feature.handler.node.NodeExecutionCompletedContext
-import ai.koog.agents.core.feature.handler.node.NodeExecutionFailedContext
-import ai.koog.agents.core.feature.handler.node.NodeExecutionStartingContext
-import ai.koog.agents.core.feature.handler.subgraph.SubgraphExecutionCompletedContext
-import ai.koog.agents.core.feature.handler.subgraph.SubgraphExecutionFailedContext
-import ai.koog.agents.core.feature.handler.subgraph.SubgraphExecutionStartingContext
-import ai.koog.serialization.TypeToken
 import kotlin.time.Clock
 
 /**
@@ -27,10 +14,9 @@ import kotlin.time.Clock
  * @property clock The clock used for time-based operations within the pipeline
  */
 public expect open class AIAgentGraphPipeline(
-    agentConfig: AIAgentConfig,
-    clock: Clock = Clock.System,
-    basePipelineDelegate: AIAgentPipelineImpl = AIAgentPipelineImpl(agentConfig, clock)
-) : AIAgentPipeline, AIAgentGraphPipelineAPI {
+    config: AIAgentConfig,
+    clock: Clock = Clock.System
+) : AIAgentGraphPipelineCommon {
 
     /**
      * Installs a feature into the pipeline with the provided configuration.
@@ -47,257 +33,4 @@ public expect open class AIAgentGraphPipeline(
         feature: AIAgentGraphFeature<TConfig, TFeatureImpl>,
         configure: TConfig.() -> Unit,
     )
-
-    //region Trigger Node Handlers
-
-    /**
-     * Notifies all registered node handlers before a node is executed.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param node The node that is about to be executed
-     * @param context The agent context in which the node is being executed
-     * @param input The input data for the node execution
-     * @param inputType The type of the input data provided to the node
-     */
-    @InternalAgentsApi
-    public open override suspend fun onNodeExecutionStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        node: AIAgentNodeBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken
-    )
-
-    /**
-     * Notifies all registered node handlers after a node has been executed.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param node The node that was executed
-     * @param context The agent context in which the node was executed
-     * @param input The input data that was provided to the node
-     * @param inputType The type of the input data provided to the node
-     * @param output The output data produced by the node execution
-     * @param outputType The type of the output data produced by the node execution
-     */
-    @InternalAgentsApi
-    public open override suspend fun onNodeExecutionCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        node: AIAgentNodeBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken,
-        output: Any?,
-        outputType: TypeToken,
-    )
-
-    /**
-     * Handles errors occurring during the execution of a node by invoking all registered node execution error handlers.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param node The instance of the node where the error occurred.
-     * @param context The context associated with the AI agent executing the node.
-     * @param input The input data provided to the node.
-     * @param inputType The type of the input data provided to the node.
-     * @param throwable The exception or error that occurred during node execution.
-     */
-    @InternalAgentsApi
-    public open override suspend fun onNodeExecutionFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        node: AIAgentNodeBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken,
-        throwable: Throwable
-    )
-
-    //endregion Trigger Node Handlers
-
-    //region Trigger Subgraph Handlers
-
-    /**
-     * Notifies all registered subgraph handlers before a subgraph is executed.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param subgraph The subgraph that is about to be executed.
-     * @param context The agent context in which the subgraph is being executed.
-     * @param input The input data for the subgraph execution.
-     * @param inputType The type of the input data provided to the subgraph.
-     */
-    @InternalAgentsApi
-    public open override suspend fun onSubgraphExecutionStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        subgraph: AIAgentSubgraphBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken
-    )
-
-    /**
-     * Notifies all registered subgraph handlers after a subgraph has been executed.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param subgraph The subgraph that was executed.
-     * @param context The agent context in which the subgraph was executed.
-     * @param input The input data provided to the subgraph.
-     * @param inputType The type of the input data provided to the subgraph.
-     * @param output The output data produced by the subgraph execution.
-     * @param outputType The type of the output data produced by the subgraph execution.
-     */
-    @InternalAgentsApi
-    public open override suspend fun onSubgraphExecutionCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        subgraph: AIAgentSubgraphBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken,
-        output: Any?,
-        outputType: TypeToken,
-    )
-
-    /**
-     * Notifies all registered subgraph handlers when a subgraph execution fails.
-     *
-     * @param eventId The unique identifier for the event group.
-     * @param executionInfo The execution information for the agent environment transformation event
-     * @param subgraph The subgraph for which the execution failed.
-     * @param context The agent context in which the subgraph execution occurred.
-     * @param input The input data that was provided to the subgraph when it failed.
-     * @param inputType The type of the input data provided to the subgraph.
-     * @param throwable The exception or error that caused the subgraph execution to fail.
-     */
-    @InternalAgentsApi
-    public open override suspend fun onSubgraphExecutionFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        subgraph: AIAgentSubgraphBase<*, *>,
-        context: AIAgentGraphContextBase,
-        input: Any?,
-        inputType: TypeToken,
-        throwable: Throwable
-    )
-
-    //endregion Trigger Subgraph Handlers
-
-    //region Interceptors
-
-    /**
-     * Intercepts node execution before it starts.
-     *
-     * @param feature The feature associated with this handler.
-     * @param handle The handler that processes before-node events
-     *
-     * Example:
-     * ```
-     * pipeline.interceptNodeExecutionStarting(feature) { eventContext ->
-     *     logger.info("Node ${eventContext.node.name} is about to execute with input: ${eventContext.input}")
-     * }
-     * ```
-     */
-    public open override fun interceptNodeExecutionStarting(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: NodeExecutionStartingContext) -> Unit
-    )
-
-    /**
-     * Intercepts node execution after it completes.
-     *
-     * @param feature The feature associated with this handler.
-     * @param handle The handler that processes after-node events
-     *
-     * Example:
-     * ```
-     * pipeline.interceptNodeExecutionCompleted(feature) { eventContext ->
-     *     logger.info("Node ${eventContext.node.name} executed with input: ${eventContext.input} and produced output: ${eventContext.output}")
-     * }
-     * ```
-     */
-    public open override fun interceptNodeExecutionCompleted(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: NodeExecutionCompletedContext) -> Unit
-    )
-
-    /**
-     * Intercepts and handles node execution errors for a given feature.
-     *
-     * @param feature The feature associated with this handler.
-     * @param handle A suspend function that processes the node execution error.
-     *
-     * Example:
-     * ```
-     * pipeline.interceptNodeExecutionFailed(feature) { eventContext ->
-     *     logger.error("Node ${eventContext.node.name} execution failed with error: ${eventContext.throwable}")
-     * }
-     * ```
-     */
-    public open override fun interceptNodeExecutionFailed(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: NodeExecutionFailedContext) -> Unit
-    )
-
-    /**
-     * Intercepts the execution of a subgraph when it starts.
-     *
-     * @param feature The graph feature associated with the AI agent for which the subgraph execution is intercepted.
-     * @param handle A suspendable lambda that handles the subgraph execution starting event context.
-     *
-     * Example:
-     * ```
-     * pipeline.interceptSubgraphExecutionStarting(feature) { eventContext ->
-     *     logger.info("Subgraph ${eventContext.subgraph.name} is about to execute with input: ${eventContext.input}")
-     * }
-     * ```
-     */
-    public open override fun interceptSubgraphExecutionStarting(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: SubgraphExecutionStartingContext) -> Unit
-    )
-
-    /**
-     * Intercepts the completion of a subgraph execution and allows handling of the event.
-     *
-     * @param feature The AI agent graph feature that specifies the feature to intercept.
-     * @param handle A suspendable function that handles the subgraph execution completion event,
-     * taking the event context as a parameter.
-     *
-     * Example:
-     * ```
-     * pipeline.interceptSubgraphExecutionCompleted(feature) { eventContext ->
-     *     logger.info("Subgraph ${eventContext.subgraph.name} executed with input: ${eventContext.input} and produced output: ${eventContext.output}")
-     * }
-     * ```
-     */
-    public open override fun interceptSubgraphExecutionCompleted(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: SubgraphExecutionCompletedContext) -> Unit
-    )
-
-    /**
-     * Intercepts and handles subgraph execution failures for a given feature.
-     *
-     * @param feature The feature associated with this handler.
-     * @param handle A suspend function that processes the subgraph execution failure event.
-     *
-     * Example:
-     * ```
-     * pipeline.interceptSubgraphExecutionFailed(feature) { eventContext ->
-     *     logger.error("Subgraph ${eventContext.subgraph.name} execution failed with error: ${eventContext.throwable}")
-     * }
-     * ```
-     */
-    public open override fun interceptSubgraphExecutionFailed(
-        feature: AIAgentGraphFeature<*, *>,
-        handle: suspend (eventContext: SubgraphExecutionFailedContext) -> Unit
-    )
-
-    //endregion Interceptors
 }
