--- agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentPipelineAPI.kt
@@ -1,420 +1,0 @@
-@file:Suppress("MissingKDocForPublicAPI")
-
-package ai.koog.agents.core.feature.pipeline
-
-import ai.koog.agents.core.agent.AIAgent
-import ai.koog.agents.core.agent.GraphAIAgent
-import ai.koog.agents.core.agent.config.AIAgentConfig
-import ai.koog.agents.core.agent.context.AIAgentContext
-import ai.koog.agents.core.agent.entity.AIAgentStorageKey
-import ai.koog.agents.core.agent.entity.AIAgentStrategy
-import ai.koog.agents.core.agent.execution.AgentExecutionInfo
-import ai.koog.agents.core.annotation.InternalAgentsApi
-import ai.koog.agents.core.environment.AIAgentEnvironment
-import ai.koog.agents.core.feature.AIAgentFeature
-import ai.koog.agents.core.feature.config.FeatureConfig
-import ai.koog.agents.core.feature.handler.agent.AgentClosingContext
-import ai.koog.agents.core.feature.handler.agent.AgentCompletedContext
-import ai.koog.agents.core.feature.handler.agent.AgentEnvironmentTransformingContext
-import ai.koog.agents.core.feature.handler.agent.AgentExecutionFailedContext
-import ai.koog.agents.core.feature.handler.agent.AgentStartingContext
-import ai.koog.agents.core.feature.handler.llm.LLMCallCompletedContext
-import ai.koog.agents.core.feature.handler.llm.LLMCallStartingContext
-import ai.koog.agents.core.feature.handler.strategy.StrategyCompletedContext
-import ai.koog.agents.core.feature.handler.strategy.StrategyStartingContext
-import ai.koog.agents.core.feature.handler.streaming.LLMStreamingCompletedContext
-import ai.koog.agents.core.feature.handler.streaming.LLMStreamingFailedContext
-import ai.koog.agents.core.feature.handler.streaming.LLMStreamingFrameReceivedContext
-import ai.koog.agents.core.feature.handler.streaming.LLMStreamingStartingContext
-import ai.koog.agents.core.feature.handler.tool.ToolCallCompletedContext
-import ai.koog.agents.core.feature.handler.tool.ToolCallFailedContext
-import ai.koog.agents.core.feature.handler.tool.ToolCallStartingContext
-import ai.koog.agents.core.feature.handler.tool.ToolValidationFailedContext
-import ai.koog.agents.core.feature.model.AIAgentError
-import ai.koog.agents.core.tools.ToolDescriptor
-import ai.koog.prompt.dsl.ModerationResult
-import ai.koog.prompt.dsl.Prompt
-import ai.koog.prompt.llm.LLModel
-import ai.koog.prompt.message.Message
-import ai.koog.prompt.streaming.StreamFrame
-import ai.koog.serialization.JSONElement
-import ai.koog.serialization.JSONObject
-import ai.koog.serialization.TypeToken
-import kotlin.reflect.KClass
-import kotlin.time.Clock
-
-/**
- * Platform-agnostic API for agent pipelines. Implemented by both the expect/actual AIAgentPipeline
- * and the shared AIAgentPipelineImpl.
- */
-public interface AIAgentPipelineAPI {
-    public val clock: Clock
-
-    public val config: AIAgentConfig
-
-    public fun <TFeature : Any> feature(
-        featureClass: KClass<TFeature>,
-        feature: AIAgentFeature<*, TFeature>
-    ): TFeature?
-
-    public fun <TConfig : FeatureConfig, TFeatureImpl : Any> install(
-        featureKey: AIAgentStorageKey<TFeatureImpl>,
-        featureConfig: TConfig,
-        featureImpl: TFeatureImpl,
-    )
-
-    public suspend fun uninstall(featureKey: AIAgentStorageKey<*>)
-
-    //region Trigger Agent Handlers
-
-    @InternalAgentsApi
-    public suspend fun <TInput, TOutput> onAgentStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        agent: AIAgent<*, *>,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onAgentCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        agentId: String,
-        runId: String,
-        result: Any?,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onAgentExecutionFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        agentId: String,
-        runId: String,
-        throwable: Throwable,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onAgentClosing(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        agentId: String
-    )
-
-    @InternalAgentsApi
-    public suspend fun onAgentEnvironmentTransforming(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        agent: GraphAIAgent<*, *>,
-        baseEnvironment: AIAgentEnvironment
-    ): AIAgentEnvironment
-
-    //endregion Trigger Agent Handlers
-
-    //region Trigger Strategy Handlers
-
-    @InternalAgentsApi
-    public suspend fun onStrategyStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        strategy: AIAgentStrategy<*, *, *>,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onStrategyCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        strategy: AIAgentStrategy<*, *, *>,
-        context: AIAgentContext,
-        result: Any?,
-        resultType: TypeToken
-    )
-
-    //endregion Trigger Strategy Handlers
-
-    //region Trigger LLM Handlers
-
-    @InternalAgentsApi
-    public suspend fun onLLMCallStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onLLMCallCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>,
-        responses: List<Message.Response>,
-        moderationResponse: ModerationResult? = null,
-        context: AIAgentContext
-    )
-
-    //endregion Trigger LLM Handlers
-
-    //region Trigger Tool Handlers
-
-    @InternalAgentsApi
-    public suspend fun onToolCallStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        toolCallId: String?,
-        toolName: String,
-        toolDescription: String?,
-        toolArgs: JSONObject,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onToolValidationFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        toolCallId: String?,
-        toolName: String,
-        toolDescription: String?,
-        toolArgs: JSONObject,
-        message: String,
-        error: AIAgentError,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onToolCallFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        toolCallId: String?,
-        toolName: String,
-        toolDescription: String?,
-        toolArgs: JSONObject,
-        message: String,
-        error: AIAgentError?,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onToolCallCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        toolCallId: String?,
-        toolName: String,
-        toolDescription: String?,
-        toolArgs: JSONObject,
-        toolResult: JSONElement?,
-        context: AIAgentContext
-    )
-
-    //endregion Trigger Tool Handlers
-
-    //region Trigger Streaming Handlers
-
-    @InternalAgentsApi
-    public suspend fun onLLMStreamingStarting(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onLLMStreamingFrameReceived(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        streamFrame: StreamFrame,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onLLMStreamingFailed(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        throwable: Throwable,
-        context: AIAgentContext
-    )
-
-    @InternalAgentsApi
-    public suspend fun onLLMStreamingCompleted(
-        eventId: String,
-        executionInfo: AgentExecutionInfo,
-        runId: String,
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>,
-        context: AIAgentContext
-    )
-
-    //endregion Trigger Streaming Handlers
-
-    //region Interceptors
-
-    public fun interceptEnvironmentCreated(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: AgentEnvironmentTransformingContext, environment: AIAgentEnvironment) -> AIAgentEnvironment
-    )
-
-    public fun interceptAgentStarting(feature: AIAgentFeature<*, *>, handle: suspend (AgentStartingContext) -> Unit)
-
-    public fun interceptAgentCompleted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: AgentCompletedContext) -> Unit
-    )
-
-    public fun interceptAgentExecutionFailed(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: AgentExecutionFailedContext) -> Unit
-    )
-
-    public fun interceptAgentClosing(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: AgentClosingContext) -> Unit
-    )
-
-    public fun interceptStrategyStarting(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (StrategyStartingContext) -> Unit
-    )
-
-    public fun interceptStrategyCompleted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (StrategyCompletedContext) -> Unit
-    )
-
-    public fun interceptLLMCallStarting(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMCallStartingContext) -> Unit
-    )
-
-    public fun interceptLLMCallCompleted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMCallCompletedContext) -> Unit
-    )
-
-    public fun interceptLLMStreamingStarting(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMStreamingStartingContext) -> Unit
-    )
-
-    public fun interceptLLMStreamingFrameReceived(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMStreamingFrameReceivedContext) -> Unit
-    )
-
-    public fun interceptLLMStreamingFailed(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMStreamingFailedContext) -> Unit
-    )
-
-    public fun interceptLLMStreamingCompleted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMStreamingCompletedContext) -> Unit
-    )
-
-    public fun interceptToolCallStarting(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallStartingContext) -> Unit
-    )
-
-    public fun interceptToolValidationFailed(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolValidationFailedContext) -> Unit
-    )
-
-    public fun interceptToolCallFailed(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallFailedContext) -> Unit
-    )
-
-    public fun interceptToolCallCompleted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallCompletedContext) -> Unit
-    )
-
-    // Short aliases
-
-    public fun interceptBeforeAgentStarted(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (AgentStartingContext) -> Unit
-    )
-
-    public fun interceptAgentFinished(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: AgentCompletedContext) -> Unit
-    )
-
-    public fun interceptAgentRunError(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (AgentExecutionFailedContext) -> Unit
-    )
-
-    public fun interceptAgentBeforeClose(feature: AIAgentFeature<*, *>, handle: suspend (AgentClosingContext) -> Unit)
-
-    public fun interceptStrategyStart(feature: AIAgentFeature<*, *>, handle: suspend (StrategyStartingContext) -> Unit)
-
-    public fun interceptStrategyFinished(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (StrategyCompletedContext) -> Unit
-    )
-
-    public fun interceptBeforeLLMCall(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMCallStartingContext) -> Unit
-    )
-
-    public fun interceptAfterLLMCall(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: LLMCallCompletedContext) -> Unit
-    )
-
-    public fun interceptToolCall(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallStartingContext) -> Unit
-    )
-
-    public fun interceptToolCallResult(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallCompletedContext) -> Unit
-    )
-
-    public fun interceptToolCallFailure(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolCallFailedContext) -> Unit
-    )
-
-    public fun interceptToolValidationError(
-        feature: AIAgentFeature<*, *>,
-        handle: suspend (eventContext: ToolValidationFailedContext) -> Unit
-    )
-
-    //endregion Interceptors
-
-    @InternalAgentsApi
-    public suspend fun prepareFeatures()
-
-    @InternalAgentsApi
-    public suspend fun closeAllFeaturesMessageProcessors()
-
-    // Note: conditional handler builders and internal helper APIs are intentionally not part of the public API
-}
