+++ agents/agents-core/src/jvmCommonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentGraphPipeline.kt
@@ -1,4 +1,4 @@
-@file:Suppress("MissingKDocForPublicAPI", "EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING", "MissingKDocForPublicAPI")
 @file:OptIn(InternalAgentsApi::class)
 
 package ai.koog.agents.core.feature.pipeline
@@ -18,11 +18,9 @@ import ai.koog.agents.core.utils.submitToMainDispatcher
 import kotlin.time.Clock
 
 public actual open class AIAgentGraphPipeline @JvmOverloads actual constructor(
-    agentConfig: AIAgentConfig,
-    clock: Clock,
-    private val basePipelineDelegate: AIAgentPipelineImpl
-) : AIAgentPipeline(agentConfig, clock),
-    AIAgentGraphPipelineAPI by AIAgentGraphPipelineImpl(agentConfig, clock, basePipelineDelegate) {
+    config: AIAgentConfig,
+    clock: Clock
+) : AIAgentGraphPipelineCommon(config, clock) {
 
     public actual fun <TConfig : FeatureConfig, TFeatureImpl : Any> install(
         feature: AIAgentGraphFeature<TConfig, TFeatureImpl>,
@@ -34,7 +32,7 @@ public actual open class AIAgentGraphPipeline @JvmOverloads actual constructor(
             pipeline = this,
         )
 
-        basePipelineDelegate.install(feature.key, featureConfig, featureImpl)
+        install(feature.key, featureConfig, featureImpl)
     }
 
     /**
