+++ agents/agents-core/src/jvmCommonMain/kotlin/ai/koog/agents/core/feature/pipeline/Interceptor.kt
