+++ agents/agents-core/src/nonJvmCommonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentGraphPipeline.kt
@@ -1,20 +1,16 @@
 @file:Suppress("MissingKDocForPublicAPI", "EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-@file:OptIn(InternalAgentsApi::class)
 
 package ai.koog.agents.core.feature.pipeline
 
 import ai.koog.agents.core.agent.config.AIAgentConfig
-import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.feature.AIAgentGraphFeature
 import ai.koog.agents.core.feature.config.FeatureConfig
 import kotlin.time.Clock
 
 public actual open class AIAgentGraphPipeline actual constructor(
-    agentConfig: AIAgentConfig,
-    clock: Clock,
-    private val basePipelineDelegate: AIAgentPipelineImpl
-) : AIAgentPipeline(agentConfig, clock),
-    AIAgentGraphPipelineAPI by AIAgentGraphPipelineImpl(agentConfig, clock, basePipelineDelegate) {
+    config: AIAgentConfig,
+    clock: Clock
+) : AIAgentGraphPipelineCommon(config, clock) {
 
     public actual fun <TConfig : FeatureConfig, TFeatureImpl : Any> install(
         feature: AIAgentGraphFeature<TConfig, TFeatureImpl>,
@@ -26,6 +22,6 @@ public actual open class AIAgentGraphPipeline actual constructor(
             pipeline = this,
         )
 
-        basePipelineDelegate.install(feature.key, featureConfig, featureImpl)
+        install(feature.key, featureConfig, featureImpl)
     }
 }
