+++ agents/agents-core/src/nonJvmCommonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentPipeline.kt
@@ -1,4 +1,4 @@
-@file:Suppress("MissingKDocForPublicAPI", "EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING", "MissingKDocForPublicAPI")
 
 package ai.koog.agents.core.feature.pipeline
 
@@ -6,6 +6,6 @@ import ai.koog.agents.core.agent.config.AIAgentConfig
 import kotlin.time.Clock
 
 public actual abstract class AIAgentPipeline actual constructor(
-    agentConfig: AIAgentConfig,
+    config: AIAgentConfig,
     clock: Clock
-) : AIAgentPipelineAPI by AIAgentPipelineImpl(agentConfig, clock)
+) : AIAgentPipelineCommon(config, clock)
