+++ agents/agents-core/src/nonJvmCommonMain/kotlin/ai/koog/agents/core/feature/pipeline/AIAgentPlannerPipeline.kt
@@ -9,9 +9,8 @@ import kotlin.time.Clock
 
 public actual open class AIAgentPlannerPipeline actual constructor(
     agentConfig: AIAgentConfig,
-    clock: Clock,
-    private val basePipelineDelegate: AIAgentPipelineImpl
-) : AIAgentPipeline(agentConfig, clock), AIAgentPlannerPipelineAPI by AIAgentPlannerPipelineImpl(agentConfig, clock, basePipelineDelegate) {
+    clock: Clock
+) : AIAgentPlannerPipelineCommon(agentConfig, clock) {
 
     /**
      * Installs a non-graph feature into the pipeline with the provided configuration.
@@ -31,6 +30,6 @@ public actual open class AIAgentPlannerPipeline actual constructor(
             pipeline = this,
         )
 
-        basePipelineDelegate.install(feature.key, featureConfig, featureImpl)
+        install(feature.key, featureConfig, featureImpl)
     }
 }
